services/device_gateway_service/requirements.txt
fastapi==0.112.2
requests==2.32.3
elasticsearch==8.15.0
python-dotenv==1.0.1
uvicorn==0.30.6
paho-mqtt==1.6.1
pika==1.3.2
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9